extracted frame.Frame mocks
Do some source re-organization: create a new file for frame mocking.
